Build a PDFViewer component using react-pdf:
This is a simple viewer with just a forward
and backward button to page throught the
PDF.  It scales reasonably on mobile and desktop.
Note that the aspect ratio is currently
hard-coded to the book size (crown quarto),
so that needs to be refactored into a property,
but works for now.